Fix PortNumber type.
ip tcp/udp ports are encoded in uint16_t format.
Use this type to prevent implicit convertion.

+++ test/test-platform/TestPlatform.cpp
@@ -38,7 +38,7 @@
 
 using std::string;
 
-CTestPlatform::CTestPlatform(const string& strClass, int iPortNumber) :
+CTestPlatform::CTestPlatform(const string& strClass, uint16_t iPortNumber) :
     mParameterMgrPlatformConnector(strClass),
     mLogger(),
     mRemoteProcessorServer(iPortNumber)
